+++ src/antares/craft/service/api_services/models/settings.py
@@ -71,6 +71,7 @@ class AdequacyPatchParametersAPI(APIBaseModel):
     threshold_initiate_curtailment_sharing_rule: int | None = None
     threshold_display_local_matching_rule_violations: int | None = None
     threshold_csr_variable_bounds_relaxation: int | None = None
+    redispatch: bool | None = None
 
     @staticmethod
     def from_user_model(user_class: AdequacyPatchParametersType) -> "AdequacyPatchParametersAPI":
@@ -103,6 +104,7 @@ class AdequacyPatchParametersAPI(APIBaseModel):
                 self.threshold_csr_variable_bounds_relaxation
             ),
             set_to_null_ntc_between_physical_out_for_first_step=self.ntc_between_physical_areas_out_adequacy_patch,
+            redispatch=self.redispatch,
         )
 
 
@@ -141,6 +143,7 @@ class AdvancedAndSeedParametersAPI(APIBaseModel):
     seed_thermal_costs: int | None = None
     seed_hydro_costs: int | None = None
     seed_initial_reservoir_levels: int | None = None
+    accurate_shave_peaks_include_short_term_storage: bool | None = None
 
     @field_validator("accuracy_on_correlation", mode="before")
     def validate_accuracy_on_correlation(cls, v: Any) -> Sequence[str] | set[str]:
@@ -171,6 +174,7 @@ class AdvancedAndSeedParametersAPI(APIBaseModel):
             renewable_generation_modelling=check_field_is_not_null(self.renewable_generation_modelling),
             accuracy_on_correlation=check_field_is_not_null(self.accuracy_on_correlation),
             initial_reservoir_levels=self.initial_reservoir_levels,
+            accurate_shave_peaks_include_short_term_storage=self.accurate_shave_peaks_include_short_term_storage,
         )
 
     def to_user_seed_parameters_model(self) -> SeedParameters: